Extract kernel code to separate module
This commit does not change any of the code statements, it simply moves
it around.
